+++ scout/server/app.py
@@ -24,7 +24,7 @@ except ImportError:
     logger.info('chanjo report not installed!')
 
 from . import extensions
-from .blueprints import public, genes, cases, login, variants, panels, pileup, dashboard, api
+from .blueprints import public, genes, cases, login, variants, panels, pileup, igv, dashboard, api
 
 
 def create_app(config_file=None, config=None):
@@ -35,7 +35,7 @@ def create_app(config_file=None, config=None):
         app.config.update(config)
     if config_file:
         app.config.from_pyfile(config_file)
-    
+
     app.config["JSON_SORT_KEYS"] = False
     current_log_level = logger.getEffectiveLevel()
     coloredlogs.install(level='DEBUG' if app.debug else current_log_level)
@@ -95,6 +95,7 @@ def register_blueprints(app):
     app.register_blueprint(pileup.pileup_bp)
     app.register_blueprint(dashboard.dashboard_bp)
     app.register_blueprint(api.api_bp)
+    app.register_blueprint(igv.igv_bp)
 
 
 def register_filters(app):